refactor: move tests to better location
