+++ src/featureMatching/descriptors/__tests__/getBriefDescriptors.test.ts
@@ -1,44 +1,6 @@
 import { ImageColorModel, Image, ImageCoordinates } from '../../../Image';
 import { getOrientedFastKeypoints } from '../../keypoints/getOrientedFastKeypoints';
-import { getHammingDistance } from '../../matching/getHammingDistance';
-import { drawKeypoints } from '../../visualize/drawKeypoints';
 import { getBriefDescriptors } from '../getBriefDescriptors';
-
-test('alphabet image, maxNbFeatures = 10', () => {
-  const image = testUtils.load('various/alphabet.jpg');
-  const grey = image.convertColor(ImageColorModel.GREY);
-
-  const keypoints = getOrientedFastKeypoints(grey, { maxNbFeatures: 10 });
-  expect(keypoints).toHaveLength(10);
-
-  let imageWithKeypoints = drawKeypoints(image, keypoints);
-
-  expect(imageWithKeypoints).toMatchImageSnapshot();
-
-  const result = getBriefDescriptors(grey, keypoints);
-
-  expect(result.descriptors).toMatchSnapshot();
-  expect(result.descriptors.length).toBe(keypoints.length);
-});
-
-test('should work with small patch size', () => {
-  const image = testUtils.load('various/alphabet.jpg');
-  const grey = image.convertColor(ImageColorModel.GREY);
-
-  const keypoints = getOrientedFastKeypoints(grey, { maxNbFeatures: 10 });
-  expect(keypoints).toHaveLength(10);
-
-  const keypoint = keypoints.slice(0, 1);
-
-  let imageWithKeypoints = drawKeypoints(image, keypoints);
-  expect(imageWithKeypoints).toMatchImageSnapshot();
-
-  const descriptor = getBriefDescriptors(grey, keypoint, {
-    patchSize: 5,
-  }).descriptors;
-
-  expect(descriptor).toMatchSnapshot();
-});
 
 test('count occurences of 1 and 0 with default options', () => {
   const image = testUtils.load('various/alphabet.jpg');
@@ -111,77 +73,4 @@ test('verify descriptor is correct (descriptorLength = 10)', () => {
   expect(descriptor).toStrictEqual(
     new Uint8Array([0, 0, 0, 0, 0, 0, 1, 0, 1, 0]),
   );
-});
-
-test.each([
-  {
-    message: 'windowSize = 7',
-    windowSize: 7,
-    expected: [
-      { srcIndex: 0, dstIndex: 0, distance: 15 },
-      { srcIndex: 0, dstIndex: 1, distance: 73 },
-      { srcIndex: 1, dstIndex: 0, distance: 77 },
-      { srcIndex: 1, dstIndex: 1, distance: 11 },
-    ],
-  },
-  {
-    message: 'windowSize = 15',
-    windowSize: 15,
-    expected: [
-      { srcIndex: 0, dstIndex: 0, distance: 13 },
-      { srcIndex: 0, dstIndex: 1, distance: 11 },
-      { srcIndex: 1, dstIndex: 0, distance: 23 },
-      { srcIndex: 1, dstIndex: 1, distance: 7 },
-    ],
-  },
-  {
-    message: 'windowSize = 31',
-    windowSize: 31,
-    expected: [
-      { srcIndex: 0, dstIndex: 0, distance: 11 },
-      { srcIndex: 0, dstIndex: 1, distance: 14 },
-      { srcIndex: 1, dstIndex: 0, distance: 20 },
-      { srcIndex: 1, dstIndex: 1, distance: 9 },
-    ],
-  },
-])('check distance for each keypoint pair ($message)', (data) => {
-  const source = testUtils
-    .load('featureMatching/polygons/scaleneTriangle.png')
-    .convertColor(ImageColorModel.GREY);
-  const sourceKeypoints = getOrientedFastKeypoints(source, {
-    windowSize: data.windowSize,
-  });
-  const sourceBrief = getBriefDescriptors(source, sourceKeypoints);
-  const destination = testUtils
-    .load('featureMatching/polygons/scaleneTriangle10.png')
-    .convertColor(ImageColorModel.GREY);
-  const destinationKeypoints = getOrientedFastKeypoints(destination, {
-    windowSize: data.windowSize,
-  });
-  const destinationBrief = getBriefDescriptors(
-    destination,
-    destinationKeypoints,
-  );
-  let result = [];
-
-  for (
-    let srcIndex = 0;
-    srcIndex < sourceBrief.descriptors.length;
-    srcIndex++
-  ) {
-    for (
-      let dstIndex = 0;
-      dstIndex < destinationBrief.descriptors.length;
-      dstIndex++
-    ) {
-      const distance = getHammingDistance(
-        sourceBrief.descriptors[srcIndex],
-        destinationBrief.descriptors[dstIndex],
-      );
-      result.push({ srcIndex, dstIndex, distance });
-    }
-  }
-
-  // 0-0 and 1-1 are the correct matches
-  expect(result).toStrictEqual(data.expected);
 });
